requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.22

# PDF Processing
pypdf>=3.17.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
Pillow>=10.0.0

# Thai OCR (Typhoon OCR)
typhoon-ocr>=0.4.0

# API Server (Production)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Ollama Integration
ollama>=0.1.6

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0

# Production Monitoring
structlog>=24.1.0
prometheus-client>=0.19.0

# Optional: OpenAI fallback
openai>=1.10.0
